// types/blogdetails.ts
import { FooterProps, HeaderProps, BlogProps } from "./home";

export interface Props {
  header: HeaderProps;
  footer: FooterProps;
  blogs: BlogProps;
  popularBlog: any;
  askquestions: any;
  sociallinks: any;
  prevblog: any;
  nextblog: any;
  prevBlogs: any;
  nextBlogs: any;
  seo: any;
  error: any;
}